requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.6.1
pydantic[email]==2.6.1
alembic==1.13.1
python-dotenv==1.0.1
uuid==1.30
pytest==8.0.1
httpx==0.26.0
pytest-asyncio==0.23.5
starlette==0.36.3
websockets==12.0
watchfiles==0.21.0
websocket-client==1.7.0
bcrypt==4.1.2
aiosmtplib==3.0.1
jinja2==3.1.3
